+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 * tiff-ios version updated to 1.0.1
 * proj4-ios verion 4.9.3
 * Bug fixes for elevation extension when handling null elevations
+* wkb version update to 1.0.7
 
 ## [1.2.1](https://github.com/ngageoint/geopackage-ios/releases/tag/1.2.1) (02-02-2017)